refactor(renderer): move the overlay layer and the section-plane frame out of index.ts
`packages/renderer/src/index.ts` was 3946 lines against the ~400-line house
rule. This takes two of the seams issue #2425 records, as a pure move: no
behaviour change, no rename of anything a caller can see.

`renderer-overlays.ts` (`RendererOverlays`, 399 lines) takes the section-plane
gizmo, the 2D section drawing / cut cap, and the standalone 3D line overlays
(annotation lines, alignment, grid, DXF, clash box and contact lines). Those
were already one unit in everything but location — created together in
`init()`, destroyed together in `destroy()`, drawn consecutively at the tail of
the render pass, fed by an upload facade that touches nothing else.

`renderer-symbolic-overlays.ts` (`SymbolicOverlays`, 160 lines) takes the
symbolic fill + text pipelines. They are a genuinely separate pair of GPU
objects sharing no state with the `Section2DOverlayRenderer` that backs the cap
AND every 3D line layer, which is why the boundary here is GPU-object
ownership rather than subject matter: "section rendering" and "standalone line
overlays" read like two topics but drive one instance, and separating them
would give one nullable, init-created / destroy-disposed object two owners in
two files. `RendererOverlays` composes this one and keeps the draw ORDER, which
is the one thing the two families share.

`render-section-plane.ts` (`resolveSectionPlaneFrame`, 306 lines) takes the
per-frame clip-plane resolution out of `renderFrame`'s encode region: the
bounds aggregation the slider is expressed in, the terrain-clip and
explicit-plane branches, and the one-shot diagnostic. The bounds come back in
the return value; the diagnostic's one-shot latch is committed through an
injected `spendLogLatch` so it still lands BEFORE the log, exactly as it did
inline — otherwise a throwing `console.info` would leave the latch unspent and
make every later section-enabled frame throw inside the encode `try`, where the
original cost one contained frame ever.

Deliberately left in `Renderer`: the render-error lifecycle. #2423's review
already established that `containFrameThrow` / `notePersistentDegradation` read
or write nine pieces of renderer state, three of them shared with code that
would stay behind — and that `consecutiveDegradedFrames`'s increment and its
reset-on-success in `render()` are one invariant, not two. Splitting that is a
different judgement call and does not belong in a mechanical move.

index.ts 3946 -> 3499 lines (-447). All three new modules are under the
~400-line rule. The exported surface is unchanged, so `pnpm api-surface:update`
produces no diff; a `patch` changeset records that consumers still receive
different built code.

Fixes #2425.
